refactor: extracting checkpoint
